refactor(google): move json into separate file
